refactor(tpu-extension-api): consolidate and clean up API surface
