+++ core/src/banking_stage/transaction_scheduler/scheduler_controller.rs
@@ -392,7 +392,7 @@ where
                 num_dropped_on_already_processed,
                 num_dropped_on_fee_payer,
                 num_dropped_on_filter_key,
-                num_dropped_on_extension_account_lock,
+                num_dropped_on_external_lock,
                 num_dropped_on_capacity,
                 num_buffered,
                 receive_time_us: _,
@@ -410,8 +410,8 @@ where
                 *num_dropped_on_already_processed;
             count_metrics.num_dropped_on_receive_fee_payer += *num_dropped_on_fee_payer;
             count_metrics.num_dropped_on_filter_key += *num_dropped_on_filter_key;
-            count_metrics.num_dropped_on_extension_account_lock +=
-                *num_dropped_on_extension_account_lock;
+            count_metrics.num_dropped_on_external_lock +=
+                *num_dropped_on_external_lock;
             count_metrics.num_dropped_on_capacity += *num_dropped_on_capacity;
             count_metrics.num_buffered += *num_buffered;
         });
