+++ core/src/banking_stage/transaction_scheduler/scheduler_metrics.rs
@@ -61,7 +61,7 @@ pub struct SchedulerCountMetricsInner {
     pub num_dropped_on_filter_key: Saturating<usize>,
     /// Number of transactions dropped because an account is write-locked by an
     /// in-flight extension batch.
-    pub num_dropped_on_extension_account_lock: Saturating<usize>,
+    pub num_dropped_on_external_lock: Saturating<usize>,
     /// Number of completed transactions received from workers.
     pub num_finished: Saturating<usize>,
     /// Number of transactions that were retryable.
@@ -129,7 +129,7 @@ impl SchedulerCountMetricsInner {
             num_unschedulable_conflicts: Saturating(num_unschedulable_conflicts),
             num_unschedulable_threads: Saturating(num_unschedulable_threads),
             num_dropped_on_filter_key: Saturating(num_dropped_on_filter_key),
-            num_dropped_on_extension_account_lock: Saturating(num_dropped_on_extension_account_lock),
+            num_dropped_on_external_lock: Saturating(num_dropped_on_external_lock),
             num_finished: Saturating(num_finished),
             num_retryable: Saturating(num_retryable),
             num_dropped_on_receive: Saturating(num_dropped_on_receive),
@@ -155,7 +155,7 @@ impl SchedulerCountMetricsInner {
             ("num_unschedulable_conflicts", num_unschedulable_conflicts, i64),
             ("num_unschedulable_threads", num_unschedulable_threads, i64),
             ("num_dropped_on_filter_key", num_dropped_on_filter_key, i64),
-            ("num_dropped_on_extension_account_lock", num_dropped_on_extension_account_lock, i64),
+            ("num_dropped_on_external_lock", num_dropped_on_external_lock, i64),
             ("num_finished", num_finished, i64),
             ("num_retryable", num_retryable, i64),
             ("num_dropped_on_receive", num_dropped_on_receive, i64),
@@ -208,7 +208,7 @@ impl SchedulerCountMetricsInner {
             || self.num_unschedulable_conflicts != Saturating(0)
             || self.num_unschedulable_threads != Saturating(0)
             || self.num_dropped_on_filter_key != Saturating(0)
-            || self.num_dropped_on_extension_account_lock != Saturating(0)
+            || self.num_dropped_on_external_lock != Saturating(0)
             || self.num_finished != Saturating(0)
             || self.num_retryable != Saturating(0)
     }
@@ -220,7 +220,7 @@ impl SchedulerCountMetricsInner {
         self.num_unschedulable_conflicts = Saturating(0);
         self.num_unschedulable_threads = Saturating(0);
         self.num_dropped_on_filter_key = Saturating(0);
-        self.num_dropped_on_extension_account_lock = Saturating(0);
+        self.num_dropped_on_external_lock = Saturating(0);
         self.num_finished = Saturating(0);
         self.num_retryable = Saturating(0);
         self.num_dropped_on_receive = Saturating(0);
