+++ tpu-extension-api-jito-solana/src/stages/bundle.rs
@@ -12,27 +12,12 @@ use {
     },
 };
 
-/// Processing stage: executes verified bundles atomically while coordinating
-/// with the packet scheduler via the gate and lock hooks.
+/// Executes verified bundles atomically.
 ///
-/// This is the last stage in the bundle pipeline and the primary reason the
-/// other extension hooks exist:
-///
-/// 1. Raises [`BundleSchedulerGate`] (`yield_flag`) so the packet scheduler
-///    pauses while a bundle is in flight — preventing interleaving.
-/// 2. Acquires [`BundleExternalLocks`] on each writable account so the
-///    scheduler's account-conflict check blocks any concurrent Agave
-///    transaction that would touch the same accounts.
-/// 3. Executes the bundle (reference stub: counts packets only).
-/// 4. Releases locks and lowers the gate so the scheduler resumes.
-///
-/// Registered via [`processing_stage`] so it is aborted *after* `BlockEngineStage`
-/// and `BundleSigverifyStage` — giving in-flight bundles a chance to finish
-/// before intake stops.
-///
-/// [`BundleSchedulerGate`]: crate::hooks::BundleSchedulerGate
-/// [`BundleExternalLocks`]: crate::hooks::BundleExternalLocks
-/// [`processing_stage`]: agave_tpu_extension_api::TpuExtensionsBuilder::processing_stage
+/// For each bundle: raises the scheduler gate, acquires write locks on the
+/// bundle's accounts, executes (reference stub: no-op), then releases locks
+/// and lowers the gate. The gate and locks together prevent the packet
+/// scheduler from interleaving regular transactions during execution.
 pub struct BundleStage {
     abort_signal: Arc<AtomicBool>,
     handle: Option<JoinHandle<()>>,
