+++ tpu-extension-api-jito-solana/src/stages/mod.rs
@@ -8,27 +8,12 @@ pub use block_engine::{BlockEngineConfig, BlockEngineStage};
 pub use bundle::BundleStage;
 pub use sigverify::BundleSigverifyStage;
 
-/// Reference-only inter-stage bundle payload.
+/// Inter-stage bundle payload (reference stand-in for the production gRPC bundle type).
 ///
-/// This type is the channel message between all three pipeline stages in this
-/// reference implementation. Production Jito-Solana derives an equivalent type
-/// from its block-engine protobuf definition.
-///
-/// In a production pipeline the pre- and post-sigverify types would differ:
-/// `BlockEngineStage` produces raw serialized packets; `BundleSigverifyStage`
-/// would produce `SanitizedTransaction` objects. The reference collapses both
-/// into a single type to keep the stub minimal.
-///
-/// `write_locks` lists the writable accounts declared by the bundle's
-/// transactions. `BundleStage` uses this list to acquire [`BundleExternalLocks`]
-/// before executing and to signal [`BundleSchedulerGate`] so the packet
-/// scheduler pauses for the duration. In production these are extracted from
-/// the transactions' writable account keys during sigverify.
-///
-/// [`BundleExternalLocks`]: crate::hooks::BundleExternalLocks
-/// [`BundleSchedulerGate`]: crate::hooks::BundleSchedulerGate
+/// Production Jito-Solana uses separate pre- and post-sigverify types; the
+/// reference collapses both into one to keep the stub minimal.
 pub(super) struct PacketBundle {
-    /// Serialized transaction bytes (pre-sigverify raw form in this reference).
+    /// Serialized transaction bytes.
     pub packets: Vec<Vec<u8>>,
     /// Writable accounts locked for the duration of bundle execution.
     pub write_locks: Vec<Pubkey>,
