+++ tpu-extension-api/src/extension/banking/mod.rs
@@ -1,9 +1,5 @@
 mod builder;
-mod config;
-mod handles;
 mod hooks;
 
 pub use builder::BankingHooksBuilder;
-pub use config::BankingConfig;
-pub use handles::BankingHandles;
-pub use hooks::BankingHooks;
+pub use hooks::{BankingConfig, BankingHooks};
