+++ tpu-extension-api/src/extension/mod.rs
@@ -1,5 +1,5 @@
 mod banking;
 mod tpu;
 
-pub use banking::{BankingConfig, BankingHandles, BankingHooks, BankingHooksBuilder};
+pub use banking::{BankingConfig, BankingHooks, BankingHooksBuilder};
 pub use tpu::{TpuExtensions, TpuExtensionsBuilder};
